+++ scripts/hooks.mjs
@@ -6,6 +6,9 @@ import {SystemManager} from "./systems/SystemManager.mjs";
 Hooks.on('init', () => {
   SystemManager.registerHandlers();
   settings.register();
+
+  if (game.modules.get('itemacro')?.active)
+    game.modules.get('itemacro').ItemMacroConfig = ItemMacroConfig;
 });
 
 Hooks.on('ready', () => {
@@ -15,6 +18,9 @@ Hooks.on('ready', () => {
   helper.register();
 });
 
+// AppV1
 Hooks.on('renderItemSheet', ItemMacroConfig._init);
+// AppV2
+Hooks.on("getHeaderControlsApplicationV2", ItemMacroConfig.getHeaderControlsApplicationV2)
 
 Hooks.on('getItemDirectoryEntryContext', (html, contextOptions) => helper.addContext(contextOptions, "Directory"));